+++ changelog.md
@@ -1,4 +1,18 @@
 # Changelog
+## 3.1.0
+- add folder structure
+- ui redesign beavers recipe compendium add responsiveness
+- add precast currency
+- add hit progression text
+- fix type selection does not change type-sub selection.
+- ui redesign recipe configuration
+- add instruction text
+- fix rescale create Item
+- fix more then two tests
+- ui redesign crafting tab on actor
+- ui redesign chatMessage
+- todo create tutorial video
+
 ## 3.0.x progress tracking and multiple tests
 ## 3.0.2 
 - fix consume on failed check